LayoutTests/http/tests/site-isolation/document-access.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<body>
<script>
    if (window.testRunner) { testRunner.dumpAsText(); testRunner.waitUntilDone() }
    onload = ()=>{ window.location = 'http://localhost:8000/site-isolation/resources/access-document.html' }
</script>
<iframe src="http://localhost:8000/site-isolation/resources/access-parent.html" frameborder=0></iframe>
</body>
